requirements.txt
fastmcp==0.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.9.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2